Change module name, refactor
