+++ src/corpus_query_language/engine.py
@@ -1,4 +1,4 @@
-import CQLEngine.functions as functions
+import corpus_query_language.functions as functions
 
 
 def parse_corpus(ast, corpus, mode, debug):
